refactor: create abstract physical tenant document isolation
have same test cases per document storage
